requirements.txt
openai
requests
moviepy
flask
flask-cors
python-dotenv
pdfminer.six
firebase-admin